chore: Install git to support installing Python packages from the Git repository
When building the Docker image of the Python environment, the installation of git is added. This step is necessary because some Python packages need to be obtained directly from the Git repository, and if git is not installed before, errors will occur when installing these packages. By pre-installing git, you can ensure that the required Python dependencies can be installed smoothly from the Git repository during the build process.